moved tests to ctl
