+++ tests/ctl/models/manual_tasks/test_manual_task.py
@@ -1,10 +1,8 @@
 from datetime import datetime, timezone
-
 import pytest
 from sqlalchemy.exc import IntegrityError
 from sqlalchemy.orm import Session
 
-from fides.api.models.connectionconfig import ConnectionConfig
 from fides.api.models.manual_tasks.manual_task import ManualTask, ManualTaskReference
 from fides.api.models.manual_tasks.manual_task_log import (
     ManualTaskLog,
@@ -15,20 +13,6 @@ from fides.api.schemas.manual_tasks.manual_task_schemas import (
     ManualTaskReferenceType,
     ManualTaskType,
 )
-
-
-@pytest.fixture
-def manual_task(db: Session, connection_config: ConnectionConfig) -> ManualTask:
-    """Create a test manual task."""
-    task = ManualTask.create(
-        db=db,
-        data={
-            "task_type": ManualTaskType.privacy_request,
-            "parent_entity_id": connection_config.id,
-            "parent_entity_type": ManualTaskParentEntityType.connection_config,
-        },
-    )
-    return task
 
 
 class TestManualTaskCreation:
